refactor(context): move core headers
