+++ include/api/bridge_internal.hpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
-#include "quantum/processor.h"
+#include "context/processor.h"
 
 #if defined(__cpp_exceptions) || defined(__EXCEPTIONS) || defined(_CPPUNWIND)
 #define SEP_HAS_EXCEPTIONS 1
